refactor(refactor-008): anti-monolith 300-line split across 14 files
Split 14 production files exceeding 300 lines into single-responsibility
modules by extracting cohesive logic into new sibling files.

- agent-core: robota → robota-base (abstract) + robota-types; execution-round
  → context/streaming/tool-results; execution-service → helpers
- agent-sessions: session → session-base (abstract) + session-components
- agent-sdk: create-session → types + runtime; agent-tool → output
- agent-command-provider: provider-command-execution → profile + setup + lifecycle + operations
- agent-plugin-logging: logging-plugin → types
- agent-provider-qwen: provider → streaming-assembly
- agent-transport-headless: headless-runner → headless-stream-json
- agent-transport-tui: useInteractiveSession → usePermissionQueue + use-interactive-session-init

All files now ≤ 300 lines. Build, typecheck, and tests pass.
