+++ packages/agent-sdk/src/tools/agent-tool.ts
@@ -28,6 +28,11 @@ import { createExecutionOriginMetadata } from '../background-tasks/index.js';
 import type { IBackgroundTaskManager } from '../background-tasks/index.js';
 import { runManagedAgentBatch } from './agent-tool-batch.js';
 import type { IAgentToolBatchJobArgs } from './agent-tool-batch.js';
+import {
+  stringifyAgentError,
+  stringifyAgentSuccess,
+  stringifyUnknownAgentType,
+} from './agent-tool-output.js';
 
 export const AGENT_TOOL_DESCRIPTION = [
   'Creates delegated subagent jobs in isolated contexts.',
@@ -188,73 +193,6 @@ function createSpawnRequest(
       label,
     }),
   };
-}
-
-function stringifyUnknownAgentType(agentType: string): string {
-  return JSON.stringify({
-    success: false,
-    mode: 'single',
-    requestedJobCount: 1,
-    startedJobCount: 0,
-    failedJobCount: 1,
-    output: '',
-    error: `Unknown agent type: ${agentType}`,
-    provenance: {
-      source: 'agent-tool-single',
-      requestedJobCount: 1,
-      startedJobCount: 0,
-      failedJobCount: 1,
-    },
-  });
-}
-
-function stringifyAgentSuccess(result: ISubagentJobResult): string {
-  const worktreePath = result.metadata?.['worktreePath'];
-  const branchName = result.metadata?.['branchName'];
-  const worktreeStatus = result.metadata?.['worktreeStatus'];
-  const worktreeNextAction = result.metadata?.['worktreeNextAction'];
-  return JSON.stringify({
-    success: true,
-    mode: 'single',
-    requestedJobCount: 1,
-    startedJobCount: 1,
-    failedJobCount: 0,
-    output: result.output,
-    agentId: result.jobId,
-    agentIds: [result.jobId],
-    provenance: {
-      source: 'agent-tool-single',
-      requestedJobCount: 1,
-      startedJobCount: 1,
-      failedJobCount: 0,
-    },
-    metadata: result.metadata,
-    ...(typeof worktreePath === 'string' ? { worktreePath } : {}),
-    ...(typeof branchName === 'string' ? { branchName } : {}),
-    ...(typeof worktreeStatus === 'string' ? { worktreeStatus } : {}),
-    ...(typeof worktreeNextAction === 'string' ? { worktreeNextAction } : {}),
-  });
-}
-
-function stringifyAgentError(message: string, agentId?: string): string {
-  const startedJobCount = agentId === undefined ? 0 : 1;
-  return JSON.stringify({
-    success: false,
-    mode: 'single',
-    requestedJobCount: 1,
-    startedJobCount,
-    failedJobCount: 1,
-    output: '',
-    error: `Sub-agent error: ${message}`,
-    agentId,
-    ...(agentId !== undefined ? { agentIds: [agentId] } : {}),
-    provenance: {
-      source: 'agent-tool-single',
-      requestedJobCount: 1,
-      startedJobCount,
-      failedJobCount: 1,
-    },
-  });
 }
 
 async function runManagedAgent(
